Moves update parent logic and spec
