+++ spec/models/batch_process_spec.rb
@@ -6,7 +6,6 @@ RSpec.describe BatchProcess, type: :model, prep_metadata_sources: true, prep_adm
   subject(:batch_process) { described_class.new }
   let(:user) { FactoryBot.create(:user, uid: "mk2525") }
   let(:csv_upload) { Rack::Test::UploadedFile.new(Rails.root.join(fixture_path, "short_fixture_ids.csv")) }
-  let(:csv_update_example) { Rack::Test::UploadedFile.new(Rails.root.join(fixture_path, "update_batch_process_example.csv")) }
   let(:csv_upload_with_source) { Rack::Test::UploadedFile.new(Rails.root.join(fixture_path, "short_fixture_ids_with_source.csv")) }
   let(:xml_upload) { Rack::Test::UploadedFile.new(Rails.root.join(fixture_path + '/goobi/metadata/30000317_20201203_140947/111860A_8394689_mets.xml')) }
   let(:xml_upload_two) { Rack::Test::UploadedFile.new(Rails.root.join(fixture_path + '/goobi/metadata/30000401_20201204_193140/IkSw55739ve_RA_mets.xml')) }
@@ -333,43 +332,6 @@ RSpec.describe BatchProcess, type: :model, prep_metadata_sources: true, prep_adm
         stub_metadata_cloud("16854285")
         stub_metadata_cloud("16172421")
         stub_metadata_cloud("30000016189097")
-      end
-
-      describe "batch update" do
-        it "includes the originating user NETid" do
-          batch_process.user_id = user.id
-          expect(batch_process.user.uid).to eq "mk2525"
-        end
-      end
-
-      context "updating a ParentObject from an import" do
-        it "can update a parent_object from a csv" do
-          expect do
-            batch_process.file = csv_upload
-            batch_process.save
-            batch_process.refresh_metadata_cloud_csv
-          end.to change { ParentObject.count }.from(0).to(5)
-
-          update_batch_process = described_class.new(batch_action: "update parent objects", user_id: user.id)
-          expect do
-            update_batch_process.file = csv_update_example
-            update_batch_process.save
-            update_batch_process.update_parent_objects
-            # byebug
-          end.not_to change { ParentObject.count }.from(5)
-          po = ParentObject.find_by(oid: 2034600)
-
-          expect(po.visibility).to eq "Public"
-          expect(po.rights_statement).to eq "The use of this image may be subject to the copyright law of the United States"
-          expect(po.extent_of_digitization).to eq "Completely digitized"
-          expect(po.digitization_note).to be_nil
-          expect(po.bib).to eq "12307100"
-          expect(po.holding).to be_nil
-          expect(po.item).to be_nil
-          expect(po.barcode).to eq "39002102340669"
-          expect(po.aspace_uri).to eq "/repositories/11/archival_objects/515305"
-          expect(po.viewing_direction).to be_nil
-        end
       end
 
       describe "batch import" do
